adicionando relacionamentos

+++ back_end/src/main/java/br/com/esoft3/warline/repository/JogadorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.esoft3.warline.modelo.Jogador;
 
-public interface JogadorRepository extends JpaRepository<Jogador, Integer>{
+public interface JogadorRepository extends JpaRepository<Jogador, Long>{
 
 }
